requirements.txt
whitenoise==3.3.1
Django==1.11.6
dj_database_url==0.4.2
djangorestframework==3.7.0
gunicorn==19.7.1
psycopg2==2.7.3.1
django-filter==1.0.4
coreapi==2.3.3
django-oauth-toolkit==1.0.0
django-import-export==0.5.1
django-cors-middleware==1.3.1
django-storages==1.6.5
django-push-notifications==1.5.0
boto3==1.4.7
django-admin-interface==0.5.8
django-rest-auth==0.9.2
pywebpush==1.5.0
django-jsonfield==1.0.1